toolkit/components/windowwatcher/test/test_blank_named_window.html

<!DOCTYPE HTML>
<html>
<!--
Test that when opening a window with the reserved name _blank that the new
window does not get that name, and that subsequent window openings with that
name result in new windows being opened.
-->
<head>
  <meta charset="utf-8">
  <title>Test named windows</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script src="head.js" type="application/javascript"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript">
    "use strict";
    add_task(function*() {
      // This magic value of 2 means that by default, when content tries
      // to open a new window, it'll actually open in a new window instead
      // of a new tab.
      yield SpecialPowers.pushPrefEnv({"set": [
        ["browser.link.open_newwindow", 2],
      ]});

      let win1 = window.open("data:text/html,<p>This is window 1 for test_blank_named_window.html</p>", "_blank");

      let name = SpecialPowers.wrap(win1)
                 .QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor)
                 .getInterface(SpecialPowers.Ci.nsIWebNavigation)
                 .QueryInterface(SpecialPowers.Ci.nsIDocShellTreeItem)
                 .name;

      is(name, "", "Should have no name");

      let win2 = window.open("data:text/html,<p>This is window 2 for test_blank_named_window.html</p>", "_blank");
      isnot(win1, win2, "Should not have gotten back the same window");

      win1.close();
      win2.close();
    });
  </script>
</body>
</html>
